w2-feat: implement client-side data validation

+++ Spring Boot/electricity-billing-system/src/main/java/com/dvgiang/electricitybillingsystem/service/ConfigurationService.java
@@ -1,16 +1,19 @@
 package com.dvgiang.electricitybillingsystem.service;
 
+import com.dvgiang.electricitybillingsystem.dto.ConfigurationDTO;
+import com.dvgiang.electricitybillingsystem.exception.NotFoundException;
 import com.dvgiang.electricitybillingsystem.model.Configuration;
 import com.dvgiang.electricitybillingsystem.repository.ConfigurationRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ConfigurationService {
     private final ConfigurationRepository configurationRepository;
 
-    //Cóntructor DI
+    //Constructor DI
     public ConfigurationService(ConfigurationRepository configurationRepository) {
         this.configurationRepository = configurationRepository;
     }
@@ -20,10 +23,15 @@ public class ConfigurationService {
     }
 
     public Configuration getConfigById(Long id) {
-        return configurationRepository.findById(id).get();
+        Optional<Configuration> configuration = configurationRepository.findById(id);
+        if (configuration.isEmpty()) {
+            throw new NotFoundException("Configuration with id = " + id + " does not exist!");
+        }
+        return configuration.get();
     }
 
-    public Configuration createConfig(Configuration configuration) {
+    public Configuration createConfig(ConfigurationDTO configurationDTO) {
+        Configuration configuration = new Configuration(configurationDTO.getName(), configurationDTO.getMinUse(), configurationDTO.getMaxUse(), configurationDTO.getPrice());
         return configurationRepository.save(configuration);
     }
 
@@ -32,6 +40,10 @@ public class ConfigurationService {
     }
 
     public String deleteConfigById(Long id) {
+        Optional<Configuration> configuration = configurationRepository.findById(id);
+        if (configuration.isEmpty()) {
+            return "Configuration ID does not exist, so not implement!";
+        }
         configurationRepository.deleteById(id);
         return "Deleted successfully!";
     }
